chore: split processing functions into files
